Cleanups
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@7549 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/plugins/jhv-3d-plugin-pfss/src/org/helioviewer/jhv/plugins/pfssplugin/PfssPluginPanel.java
@@ -40,7 +40,7 @@ public class PfssPluginPanel extends SmallPanel {
         levelSpinner.addChangeListener(new ChangeListener() {
             @Override
             public void stateChanged(ChangeEvent e) {
-                PfssSettings.qualityReduction = 8 - ((Integer) levelSpinner.getValue()).intValue();
+                PfssSettings.qualityReduction = 8 - (Integer) levelSpinner.getValue();
                 Displayer.display();
             }
         });
